requirements.txt
customtkinter
CTkMessagebox
pygame
urx
math3d
opencv-python
numpy
scipy
pyserial
robodk
